+++ src/App.js
@@ -98,7 +98,11 @@ enableScreens();
 const containerStyle = { flex: 1 };
 
 class OldApp extends Component {
-  state = { appState: AppState.currentState, initialRoute: null };
+  state = {
+    appState: AppState.currentState,
+    initialRoute: null,
+    eventSubscription: null,
+  };
 
   /**
    * There's a race condition in Branch's RN SDK. From a cold start, Branch
@@ -143,7 +147,11 @@ class OldApp extends Component {
     InteractionManager.runAfterInteractions(() => {
       rainbowTokenList.update();
     });
-    AppState?.addEventListener('change', this?.handleAppStateChange);
+    const eventSub = AppState?.addEventListener(
+      'change',
+      this?.handleAppStateChange
+    );
+    this.setState({ eventSubscription: eventSub });
     rainbowTokenList.on('update', this.handleTokenListUpdate);
     appEvents.on('transactionConfirmed', this.handleTransactionConfirmed);
 
@@ -183,9 +191,9 @@ class OldApp extends Component {
   }
 
   componentWillUnmount() {
-    AppState.removeEventListener('change', this.handleAppStateChange);
-    rainbowTokenList?.off?.('update', this.handleTokenListUpdate);
-    this.branchListener?.();
+    this.state.eventSubscription.remove();
+    rainbowTokenList.off('update', this.handleTokenListUpdate);
+    this.branchListener();
   }
 
   identifyFlow = async () => {